+++ src/components/TimesSquarePage/index.js
@@ -1,0 +1,2 @@
+export * from './TimesSquarePage';
+export { default } from './TimesSquarePage';
