--- src/components/TimesSquareViewer/index.js
@@ -1,2 +1,0 @@
-export * from './TimesSquareViewer';
-export { default } from './TimesSquareViewer';
